fix(backend): satisfy mypy --strict on geodetic + alerts
* `geodetic.gmst_radians`: cast the modulo result to `float` so mypy
  stops inferring `Any` from the `numpy`-tainted return path.
* `alerts._discord_embed`: the conjunction row payload is typed as
  `dict[str, object]` so `float(value)` cannot be statically proven
  safe; the keys really are floats by construction at call sites,
  guard the three coercions with localised `type: ignore[arg-type]`
  comments.

`mypy src` -- "no issues found in 54 source files".

+++ backend/src/oc/infrastructure/propagation/geodetic.py
@@ -67,7 +67,7 @@ def gmst_radians(when: datetime) -> float:
         - 6.2e-6 * t * t * t
     )
     gmst_rad = (gmst_seconds % 86400.0) / 240.0 * math.pi / 180.0
-    return gmst_rad % (2.0 * math.pi)
+    return float(gmst_rad % (2.0 * math.pi))
 
 
 def teme_to_ecef(position_teme_km: np.ndarray, when: datetime) -> np.ndarray:
